Move all interfaces for services to its own namespace and folder
