+++ CICD/App.cs
@@ -9,6 +9,7 @@ using CICDSystem.Factories;
 using CICDSystem.Reactables;
 using CICDSystem.Reactables.Core;
 using CICDSystem.Services;
+using CICDSystem.Services.Interfaces;
 using Nuke.Common.ProjectModel;
 using SimpleInjector;
