+++ CICD/Factories/HttpClientFactory.cs
@@ -6,7 +6,7 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using CICDSystem.Guards;
 using CICDSystem.Reactables.Core;
-using CICDSystem.Services;
+using CICDSystem.Services.Interfaces;
 using Octokit;
 using Octokit.Internal;
