+++ CICD/Services/BranchValidatorService.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using CICDSystem.Factories;
 using CICDSystem.Guards;
 using CICDSystem.Reactables.Core;
+using CICDSystem.Services.Interfaces;
 using Octokit;
 
 namespace CICDSystem.Services;
