+++ CICD/Services/Interfaces/IGitHubTokenService.cs
@@ -2,7 +2,7 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace CICDSystem.Services;
+namespace CICDSystem.Services.Interfaces;
 
 /// <summary>
 /// Creates tokens depending whether or not the execution context is local or on the server.
