+++ CICD/Services/Interfaces/IWorkflowService.cs
@@ -5,7 +5,7 @@
 using System;
 using System.IO;
 
-namespace CICDSystem.Services;
+namespace CICDSystem.Services.Interfaces;
 
 /// <summary>
 /// Generates workflow files for the targets in the Nuke build system.
