+++ CICD/Services/SecretService.cs
@@ -8,6 +8,7 @@ using System.Collections.Generic;
 using System.IO.Abstractions;
 using System.Linq;
 using CICDSystem.Guards;
+using CICDSystem.Services.Interfaces;
 
 namespace CICDSystem.Services;
