+++ Testing/CICDTests/Services/BranchValidatorServiceTests.cs
@@ -7,6 +7,7 @@ using CICDSystem;
 using CICDSystem.Factories;
 using CICDSystem.Reactables.Core;
 using CICDSystem.Services;
+using CICDSystem.Services.Interfaces;
 using CICDSystemTests.Helpers;
 using FluentAssertions;
 using Moq;
