+++ Testing/CICDTests/Services/GitHubTokenServiceTests.cs
@@ -3,6 +3,7 @@
 // </copyright>
 
 using CICDSystem.Services;
+using CICDSystem.Services.Interfaces;
 using FluentAssertions;
 using Moq;
 using Xunit;
