+++ Testing/CICDTests/Services/PullRequestServiceTests.cs
@@ -6,6 +6,7 @@ using System.Net;
 using CICDSystem.Factories;
 using CICDSystem.Reactables.Core;
 using CICDSystem.Services;
+using CICDSystem.Services.Interfaces;
 using CICDSystemTests.Helpers;
 using FluentAssertions;
 using Moq;
